requirements/requirements.txt
fastapi==0.105.0
uvicorn==0.29.0
openai==1.12.0
pyyaml==6.0.1
pathlib==1.0.1
numpy==1.26.4
python-dotenv==1.0.1